# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sessions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mengembalikan daftar sesi perebusan pengguna",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BoisingSessions"
                ],
                "summary": "GetSessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BoilingSessionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Membuat sesi perebusan baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BoisingSessions"
                ],
                "summary": "CreateSession",
                "parameters": [
                    {
                        "description": "Data sesi perebusan",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BoilingSessionCreation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.BoilingSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{public_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mengambil detail sesi perebusan berdasarkan public id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BoisingSessions"
                ],
                "summary": "GetSessionByPublicID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID Sesi",
                        "name": "public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BoilingSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{public_id}/finish": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menyelesaikan sesi perebusan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BoisingSessions"
                ],
                "summary": "FinishSession",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID Sesi",
                        "name": "public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{public_id}/records": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mengemambil record telemetry untuk session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BoisingSessions"
                ],
                "summary": "GetSessionRecords",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID Sesi",
                        "name": "public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/utils.Response"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{public_id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Memperbarui status sesi perebusan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BoisingSessions"
                ],
                "summary": "UpdateSessionStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID Sesi",
                        "name": "public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status baru sesi",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BoilingSessionStatusUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{session_id}/records": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mengambil semua records berdasarkan session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SessionRecords"
                ],
                "summary": "GetRecords",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID session",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SessionRecordOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{session_id}/records": {
            "post": {
                "description": "Endpoint pengujian untuk simulasi data sensor dari ESP. Hapus sebelum production.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SessionRecords"
                ],
                "summary": "[TEST ONLY] CreateRecord",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID session",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data sensor",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SessionRecordInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.BoilingSessionCreation": {
            "type": "object",
            "required": [
                "esp_public_id",
                "fabric_type",
                "kompor_public_id"
            ],
            "properties": {
                "esp_public_id": {
                    "type": "string"
                },
                "fabric_type": {
                    "type": "string"
                },
                "kompor_public_id": {
                    "type": "string"
                }
            }
        },
        "models.BoilingSessionResponse": {
            "type": "object",
            "properties": {
                "boiling_status": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "esp_id": {
                    "type": "integer"
                },
                "fabric_type": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "kompor_id": {
                    "type": "integer"
                },
                "public_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.BoilingSessionStatusUpdate": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.SessionRecordInput": {
            "type": "object",
            "properties": {
                "air_temp": {
                    "type": "number"
                },
                "humidity": {
                    "type": "number"
                },
                "session_pub_id": {
                    "type": "string"
                },
                "water_temp": {
                    "type": "number"
                }
            }
        },
        "models.SessionRecordOutput": {
            "type": "object",
            "properties": {
                "air_temp": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "humidity": {
                    "type": "number"
                },
                "public_id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "integer"
                },
                "water_temp": {
                    "type": "number"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}